Rename routes/agents/types.ts to shared.ts
The module contains runtime constants and helpers alongside types, so
the name types.ts was misleading. Rename to shared.ts for clarity.
