+++ apps/server/src/routes/agents/index.ts
@@ -1,13 +1,13 @@
 import type { FastifyInstance } from "fastify";
 
-import type { AgentRouteDeps } from "./types.js";
+import type { AgentRouteDeps } from "./shared.js";
 import { registerAgentCrudRoutes } from "./crud-routes.js";
 import { registerAgentEventRoutes } from "./events-routes.js";
 import { registerAgentLifecycleRoutes } from "./lifecycle-routes.js";
 import { registerAgentStreamingRoutes } from "./streaming-routes.js";
 import { registerAgentTerminalRoutes } from "./terminal-routes.js";
 
-export type { AgentRouteDeps } from "./types.js";
+export type { AgentRouteDeps } from "./shared.js";
 
 export async function registerAgentRoutes(
   app: FastifyInstance,
